refactor: move sync/chain checks to their own files
